# fix: grouping audio frames causes memory problems and is maybe slightly slower?