@IvanovCosmin 2024-05-17
I am unsure if the gc call is still necessary.
It seems to be related to the sampler, but also it seems to make no difference now.

I'll let you decide