@trungkienbkhn 2024-05-20
@IvanovCosmin , hello. I believe that memory will not increase over time in the _group_frames function because memory is released after calling fifo.read().
```python
def _group_frames(frames, num_samples=None):
    fifo = av.audio.fifo.AudioFifo()

    for frame in frames:
        frame.pts = None  # Ignore timestamp check.
        fifo.write(frame)

        if num_samples is not None and fifo.samples >= num_samples:
            yield fifo.read()

    if fifo.samples > 0:
        yield fifo.read()
```
Additionally, I think that grouping frames into a consistent size (500,000) helps standardize the processing pipeline, making the next steps (resampling, ...) more stable and avoiding errors with long audio inputs