@IvanovCosmin 2024-05-20
> believe that memory will not increase over time in the _group_frames function because memory is released after calling fifo.read().

It does not increase over time, it just increases too much for too long.

I think that AV does something weird here. I follow your judgement, but this is not what happens in practice and it does not work well in production envs. 

1. Memory is not released when it should, we can see the increase persisting for longer than the scope of the  decode function, when the generator should be released. It seems to be enventually released, but long after it is supposed to.
2. It is unjustifiable to allocate more memory than the actual video consumes. My 8 second video is kilobytes, and it requires 18 MB to do just this step. For 3 minute videos we could see 60-80 MBs regularely allcoated. Let's now extrapolate to hour long videos. 
3. I did not find any issues with stability after removing that.